collapsed timeline componentize
